Saluki hack: Remove the hsm_start logic from boot
We have already harts waiting in wfi in bootloader before they even enter
to opensbi.

Now the domain init just first fires up all the SBI harts running from
the bootloader and later they get stuck in waiting for the IPI inside opensbi

Removing this logic lets us control which harts are started in our bootloader
code and when this happens.

TODO: It should be designed how to handle this situation and make some proper fix

Signed-off-by: Jukka Laitinen <jukkax@ssrc.tii.ae>